MOM-128 Started reworking the back end for the disposition. Removed some tables from the db and merged the disposition class into publication.
