--- src/main/java/no/dusken/momus/service/repository/DispositionRepository.java
@@ -1,9 +1,0 @@
-package no.dusken.momus.service.repository;
-
-import no.dusken.momus.model.Disposition;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface DispositionRepository extends JpaRepository<Disposition, Long> {
-
-}
