+++ src/main/java/no/dusken/momus/service/repository/PublicationRepository.java
@@ -21,4 +21,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PublicationRepository extends JpaRepository<Publication, Long> {
 
+
 }
